refactor: move chart to own component
