+++ libs/i18n/src/resources/locales/ar/playground/room.json
@@ -5,8 +5,7 @@
 		"description": "تحديث إعدادات الغرفة وتعديل سلوك الدردشة",
 		"open": "فتح إعدادات الغرفة",
 		"close": "إغلاق إعدادات الغرفة",
-		"edit": "تعديل إعدادات الغرفة",
-		"compactTooltip": "اضغط محادثتك لتوفير الرموز. تبقى رسائلك السابقة كما هي."
+		"edit": "تعديل إعدادات الغرفة"
 	},
 	"form": {
 		"modelLabel": "النموذج",
