+++ libs/i18n/src/resources/locales/fr/playground/room.json
@@ -5,8 +5,7 @@
 		"description": "Mettez à jour les paramètres de la salle et modifiez le comportement du chat",
 		"open": "Ouvrir les paramètres de la salle",
 		"close": "Fermer les paramètres de la salle",
-		"edit": "Modifier les paramètres de la salle",
-		"compactTooltip": "Compactez votre conversation pour économiser des jetons. Vos messages précédents restent inchangés."
+		"edit": "Modifier les paramètres de la salle"
 	},
 	"form": {
 		"modelLabel": "Modèle",
