+++ libs/i18n/src/resources/locales/hi/playground/room.json
@@ -5,8 +5,7 @@
 		"description": "रूम सेटिंग्स अपडेट करें और चैट का व्यवहार संशोधित करें",
 		"open": "रूम सेटिंग्स खोलें",
 		"close": "रूम सेटिंग्स बंद करें",
-		"edit": "रूम सेटिंग्स संपादित करें",
-		"compactTooltip": "टोकन बचाने के लिए अपनी बातचीत को संकुचित करें। आपके पिछले संदेश वैसे ही रहेंगे।"
+		"edit": "रूम सेटिंग्स संपादित करें"
 	},
 	"form": {
 		"modelLabel": "मॉडल",
