+++ libs/i18n/src/resources/locales/ja/playground/room.json
@@ -5,8 +5,7 @@
 		"description": "ルーム設定を更新し、チャットの動作を変更します",
 		"open": "ルーム設定を開く",
 		"close": "ルーム設定を閉じる",
-		"edit": "ルーム設定を編集",
-		"compactTooltip": "会話を圧縮してトークンを節約します。これまでのメッセージはそのまま残ります。"
+		"edit": "ルーム設定を編集"
 	},
 	"form": {
 		"modelLabel": "モデル",
