+++ libs/shared/src/components/engine/index.ts
@@ -1,3 +1,2 @@
-export * from "./context-usage-indicator";
 export * from "./engine-select";
 export * from "./new-engine-input";
